requirements.txt
wheel
pyaml
setproctitle
sqlalchemy
sqlalchemy-pysqlite3
pyvcloud
pyvmomi
hcloud
boto3
google-api-python-client
google-auth
google-cloud-datastore
azure-common
azure-mgmt-sql
azure-mgmt-resource
azure-mgmt-rdbms
azure-identity
mysql
mysql-connector
hetzner
proxmoxer
prometheus-client
psutil
sentry-sdk
apache-libcloud